# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VelozTrade API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /social/posts:
    get:
      operationId: listSocialPosts
      tags: [social]
      summary: List own social posts
      description: Returns the authenticated user's own trade idea posts
      responses:
        "200":
          description: List of social posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialPost"
    post:
      operationId: createSocialPost
      tags: [social]
      summary: Create a trade idea post
      description: Publishes a new trade idea to the social feed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialPostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialPost"
...
  /positions:
    get:
      operationId: listPositions
      tags: [positions]
      summary: List open positions
      description: Returns all currently open trading positions
      responses:
        "200":
          description: Open positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"
    post:
      operationId: openPosition
      tags: [positions]
      summary: Open a new position
      description: Places a new buy or sell trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PositionInput"
      responses:
        "201":
          description: Position opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
        "400":
          description: Invalid trade parameters
  /positions/{id}:
    get:
      operationId: getPosition
      tags: [positions]
      summary: Get position details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Position details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
        "404":
          description: Position not found
    patch:
      operationId: updatePosition
      tags: [positions]
      summary: Modify position SL/TP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PositionUpdate"
      responses:
        "200":
          description: Position updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
        "404":
          description: Position not found
    delete:
      operationId: closePosition
      tags: [positions]
      summary: Close a position
      description: Closes an open trading position. Pass the closePrice the client observed to ensure zero-slippage execution.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClosePositionInput"
      responses:
        "200":
          description: Position closed, returns closed order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Position not found
...
components:
  schemas:
...
    Position:
      type: object
      required: [id, symbol, type, direction, volume, openPrice, currentPrice, profit, swap, openTime, lotSize]
      properties:
        id:
          type: integer
        symbol:
          type: string
        symbolName:
          type: string
        type:
          type: string
          enum: [market]
        direction:
          type: string
          enum: [buy, sell]
        volume:
          type: number
        lotSize:
          type: number
        openPrice:
          type: number
        currentPrice:
          type: number
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        profit:
          type: number
        swap:
          type: number
        commission:
          type: number
        openTime:
          type: string
    PositionInput:
      type: object
      required: [symbol, direction, volume]
      properties:
        symbol:
          type: string
        direction:
          type: string
          enum: [buy, sell]
        volume:
          type: number
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
    PositionUpdate:
      type: object
      properties:
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
    Order:
      type: object
      required: [id, symbol, direction, volume, openPrice, closePrice, profit, openTime, closeTime]
      properties:
        id:
          type: integer
        symbol:
          type: string
        symbolName:
          type: string
        direction:
          type: string
          enum: [buy, sell]
        volume:
          type: number
        openPrice:
          type: number
        closePrice:
          type: number
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        profit:
          type: number
        swap:
          type: number
        commission:
          type: number
        openTime:
          type: string
        closeTime:
          type: string
        journalNote:
          type: ["string", "null"]
        strategyTag:
          type: ["string", "null"]
        sentimentRating:
          type: ["integer", "null"]
...
    SocialPost:
      type: object
      required: [id, clerkUserId, content, createdAt]
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        content:
          type: string
        symbol:
          type: ["string", "null"]
        createdAt:
          type: string
    ClosePositionInput:
      type: object
      properties:
        closePrice:
          type: number
          description: The bid/ask price the client saw when the user clicked Close. Used for zero-slippage execution.
    SocialPostInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        symbol:
          type: ["string", "null"]